Add simulation path logic and commands.
Add a 'simulate on <path>' command, which declares the given directory
as the source of exposure files.

Files in the directory are returned in sorted order. As it stands we
loop back to the beginning when all the files have been read.

'simulate off' goes back to reading from the camera.