Slightly improve doExpose handling of gen2 failures.
This is not great. Basically, gen2.getVisit call for a gen2 frame
number and then inserts a row into tel_status. Both of those can
timeout, but only the getFrame is fatal. In most cases, 'mcs doExpose'
is given a visit, in which case we can always keep going.

We thread the needle, only failing the exposure when we cannot tell what
the problem is.